* CHG:  Callback app id type

+++ models/mercadolibre/Callback.go
@@ -6,9 +6,9 @@ type MercadoCallback struct {
 	ID            int       `json:"id"`
 	MercadoID     string    `json:"_id"`
 	Resource      string    `json:"resource"`
-	UserID        int       `json:"user_id"`
+	UserID        int64     `json:"user_id"`
 	Topic         string    `json:"topic"`
-	ApplicationID int       `json:"application_id"`
+	ApplicationID int64     `json:"application_id"`
 	Attempts      int       `json:"attempts"`
 	Sent          time.Time `json:"sent"`
 	Received      time.Time `json:"received"`
